rename: 컴포넌트가 사용되는 디렉토리로 이동
- create, join -> chatList

#42
