+++ src/pages/auth/chat/chatList/ChatList.tsx
@@ -3,7 +3,7 @@ import { useNavigate } from "react-router-dom";
 import { getAuth, isAuth } from "userAuth";
 import ChatItem from "./ChatItem";
 import * as S from "./style";
-import CreateChatRoom from "../createChatRoom/CreateChatRoom";
+import CreateChatRoom from "./create/CreateBtn";
 import { ChatListType } from "socket/chat";
 
 type propsType = {
